Conflicts: 	Jenkinsfile 	examples/pytorch/gcn/gcn.py 	examples/pytorch/gcn/gcn_spmv.py 	python/dgl/graph.py 	python/dgl/graph_index.py 	src/graph/graph.cc 	src/graph/graph_op.cc

+++ src/c_api_common.cc
@@ -24,7 +24,7 @@ DLManagedTensor* CreateTmpDLManagedTensor(const TVMArgValue& arg) {
 
 PackedFunc ConvertNDArrayVectorToPackedFunc(const std::vector<NDArray>& vec) {
     auto body = [vec](TVMArgs args, TVMRetValue* rv) {
-        size_t which = args[0];
+        int which = args[0];
         if (which >= vec.size()) {
             LOG(FATAL) << "invalid choice";
         } else {
